feat : GET /related-to 요청 처리
- 실제 DB 접근 하도록 변경
- controller/post.js 생성
- post id 가 유효 하지 않아도 204코드로 변경
- swagger 변경
