+++ swagger/swagger-api.yaml
@@ -114,9 +114,7 @@ paths:
         204:
           description: "No Contents.\n해당 장소에 다른 게시글이 없을 때\npage number 가 커서 더이상 글이 없을 때"
         400:
-          description: "Bad Request.\n요청에 post id or page number 가 없을 때"
-        404:
-          description: "Not Found.\n존재 하지 않는 post id 일 때"
+          description: "Bad Request.\n요청에 post id or page number 가 없을 때\n존재 하지 않는 post id 일 때"
 definitions:
   fullPostInfo:
     type: "object"
